requirements.txt
boto3>=1.34
google-cloud-storage>=2.16
tenacity>=8.2
pytest>=8.0
